requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
sentence-transformers
streamlit
python-dotenv
groq
langchain-groq
openai